feature/Secure endpoints v4.0.0 related to balance
